real_esrgan/requirements.txt
onnxruntime-gpu
Pillow
numpy
opencv-python==4.5.2.54
basicsr
click
gdown